pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "climatevis"
version = "0.1.0"
description = "Professional Weather Data Visualization Library"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "ClimateVis Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Visualization",
]
keywords = ["climate", "weather", "visualization", "plotly", "meteorology"]
requires-python = ">=3.8"
dependencies = [
    "plotly>=5.0",
    "pandas>=1.3",
    "numpy>=1.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
marimo = ["marimo"]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8"
]

[project.urls]
Homepage = "https://github.com/climatevis/climatevis"
Documentation = "https://github.com/climatevis/climatevis#readme"
Repository = "https://github.com/climatevis/climatevis.git"
Issues = "https://github.com/climatevis/climatevis/issues"

# Package configuration
[tool.setuptools]
packages = ["climatevis", "climatevis.plots", "climatevis.util", "climatevis.components"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"climatevis.templates" = ["*.yaml", "*.yml"]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88